update rebalance logic to use only big.Int

+++ packages/go-kosu/abci/tags.go
@@ -44,11 +44,11 @@ func NewRoundInfoFromTags(tags map[string]string) (*types.RoundInfo, error) {
 }
 
 // NewTagsFromOrderInfo returns the KVPairs necessary to make up Order transaction tags
-func NewTagsFromOrderInfo(orderID []byte, posterAddress store.Address, newLimit uint32) []common.KVPair {
+func NewTagsFromOrderInfo(orderID []byte, posterAddress store.Address, newLimit uint64) []common.KVPair {
 	return []common.KVPair{
 		{Key: []byte("tx.type"), Value: []byte("order")},
 		{Key: []byte("order.id"), Value: orderID},
 		{Key: []byte("order.poster"), Value: posterAddress.Bytes()},
-		{Key: []byte("poster.limit"), Value: []byte(strconv.FormatUint(uint64(newLimit), 10))},
+		{Key: []byte("poster.limit"), Value: []byte(strconv.FormatUint(newLimit, 10))},
 	}
 }
